+++ src/main/java/domain/piece/piecerole/Knight.java
@@ -9,7 +9,7 @@ import static domain.game.Direction.RIGHT_UP;
 import static domain.game.Direction.UP_LEFT;
 import static domain.game.Direction.UP_RIGHT;
 
-import domain.piece.Movable;
+import domain.game.Movable;
 import domain.position.Position;
 import java.util.List;
 import java.util.Objects;
