--- src/test/java/domain/game/PieceFixture.java
@@ -1,7 +1,0 @@
-package domain.game;
-
-public final class PieceFixture {
-    // PieceMover
-
-    //
-}
