+++ src/test/java/fixture/PositionFixture.java
@@ -1,4 +1,4 @@
-package domain.game;
+package fixture;
 
 import domain.position.File;
 import domain.position.Position;
@@ -8,18 +8,26 @@ public final class PositionFixture {
     public static Position generateC1Position() {
         return new Position(new File('c'), new Rank(1));
     }
+
     public static Position generateC2Position() {
         return new Position(new File('c'), new Rank(2));
     }
+
     public static Position generateC3Position() {
         return new Position(new File('c'), new Rank(3));
     }
+
     public static Position generateB1Position() {
         return new Position(new File('b'), new Rank(1));
     }
-    public static Position generateB2Position() {return new Position(new File('b'), new Rank(2));}
-    public static Position generateB7Position() {return new Position(new File('b'), new Rank(7));}
 
+    public static Position generateB2Position() {
+        return new Position(new File('b'), new Rank(2));
+    }
+
+    public static Position generateB7Position() {
+        return new Position(new File('b'), new Rank(7));
+    }
 
 
 }
